+++ app/src/src_nonfreenet/org/breezyweather/sources/china/json/ChinaForecastHourly.kt
@@ -21,7 +21,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ChinaForecastHourly(
     val desc: String?,
-    val temperature: ChinaForecastHourlyTemperature?,
+    val temperature: ChinaValueListInt?,
     val weather: ChinaValueListInt?,
     val wind: ChinaHourlyWind?
 )